[WIP] Refactor loader / Implement loader screen fails to compile
