+++ src/script/player/renderer/renderer.js
@@ -32,9 +32,10 @@ export class Renderer {
       RenderUtil.strokeCircle(g, 0, 0, 160, this.colorScheme.lane[2], 1)
       RenderUtil.strokeCircle(g, 0, 0, 190, this.colorScheme.lane[3], 1)
 
-      RenderUtil.strokeArc(g, 0, 0, 130, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.promiseBmsonProgress, this.colorScheme.note.long.active[1], 5, false)
-      RenderUtil.strokeArc(g, 0, 0, 160, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.promiseChartProgress, this.colorScheme.note.long.active[2], 5, false)
-      RenderUtil.strokeArc(g, 0, 0, 190, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.promiseAssetProgress, this.colorScheme.note.long.active[3], 5, false)
+      if(player.loaderBmson) RenderUtil.strokeArc(g, 0, 0, 100, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.loaderBmson.progress, this.colorScheme.note.long.active[0], 5, false)
+      if(player.loaderChart) RenderUtil.strokeArc(g, 0, 0, 130, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.loaderChart.progress, this.colorScheme.note.long.active[1], 5, false)
+      if(player.loaderAsset) RenderUtil.strokeArc(g, 0, 0, 160, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.loaderAsset.progress, this.colorScheme.note.long.active[2], 5, false)
+      if(player.loaderAudio) RenderUtil.strokeArc(g, 0, 0, 190, -Math.PI / 2, -Math.PI / 2 + Math.PI * 2 * player.loaderAudio.progress, this.colorScheme.note.long.active[3], 5, false)
       g.restore()
 
       RenderUtil.fillRect(g, 0, 225, 800, 150, "rgba(255, 255, 255, 0.6)")
